crud mahasiswa done

+++ routes/web.php
@@ -35,23 +35,33 @@ Route::post('/logout', [AuthenticatedSessionController::class, 'destroy'])->name
 
 Route::get('/masterdata', [DashboardController::class, 'index'])->name('admin.index');
 
-// Dosen
-Route::get('/daftar-dosen', [DosenController::class, 'index'])->name('admin.dosen.index');
+Route::prefix('admin')->name('admin.')->group(function () {
+    Route::get('/dosen', [DosenController::class, 'index'])->name('dosen.index');
+    Route::get('/dosen/create', [DosenController::class, 'create'])->name('dosen.create');
+    Route::post('/dosen', [DosenController::class, 'store'])->name('dosen.store');
+    Route::get('/dosen/{id}/edit', [DosenController::class, 'edit'])->name('dosen.edit');
+    Route::put('/dosen/{id}', [DosenController::class, 'update'])->name('dosen.update');
+    Route::delete('/dosen/{id}', [DosenController::class, 'destroy'])->name('dosen.destroy');
+});
 
-// mahasiswa
-Route::get('/daftar-mahasiswa', [MahasiswaController::class, 'index'])->name('admin.mahasiswa.index');
-Route::get('/mahasiswa/create', [MahasiswaController::class, 'create'])->name('mahasiswa.create');
-Route::get('/mahasiswa/{id}/edit', [MahasiswaController::class, 'edit'])->name('mahasiswa.edit');
-Route::put('/mahasiswa/{id}', [MahasiswaController::class, 'update'])->name('mahasiswa.update');
-Route::delete('/mahasiswa/{id}', [MahasiswaController::class, 'destroy'])->name('mahasiswa.destroy');
-
-
-// Route::group(['prefix' => 'admin', 'middleware' => 'auth'], function () {
-//     Route::get('/', 'App\Http\Controllers\Admin\DashboardController@index')->name('admin.dashboard');
-//     Route::resource('mahasiswa', 'App\Http\Controllers\Admin\MahasiswaController');
-//     Route::resource('dosen', 'App\Http\Controllers\Admin\DosenController');
+// Route::prefix('admin')->group(function () {
+//     Route::get('/dosen', [DosenController::class, 'index'])->name('admin.dosen.index');
+//     Route::get('/dosen/create', [DosenController::class, 'create'])->name('admin.dosen.create');
+//     Route::post('/dosen', [DosenController::class, 'store'])->name('admin.dosen.store');
+//     Route::get('/dosen/{id}/edit', [DosenController::class, 'edit'])->name('admin.dosen.edit');
+//     Route::post('/dosen/{id}', [DosenController::class, 'update'])->name('admin.dosen.update');
+//     Route::delete('/dosen/{id}', [DosenController::class, 'destroy'])->name('admin.dosen.destroy');
 // });
 
+// Mahasiswa
+Route::prefix('admin')->group(function () {
+    Route::get('/mahasiswa', [MahasiswaController::class, 'index'])->name('admin.mahasiswa.index');
+    Route::get('/mahasiswa/create', [MahasiswaController::class, 'create'])->name('admin.m.create');
+    Route::post('/mahasiswa', [MahasiswaController::class, 'store'])->name('admin.mahasiswa.store');
+    Route::get('/mahasiswa/{id}/edit', [MahasiswaController::class, 'edit'])->name('admin.mahasiswa.edit');
+    Route::post('/mahasiswa/{id}', [MahasiswaController::class, 'update'])->name('admin.mahasiswa.update');
+    Route::delete('/mahasiswa/{id}', [MahasiswaController::class, 'destroy'])->name('admin.mahasiswa.destroy');
+});
 
 
 require __DIR__.'/auth.php';
